#49 - OSGi-friendly KaiZen Parser: add explicit dependencies to Jackson libs
We are currently using RepreZen Bundles plugin to contain third-party
non-Tycho dependencies. This is first change in the effort to replace
the bundles with Eclipse Orbit plugins.

Note that new YAMLMapper() was replaced by new ObjectMapper(new
YAMLFactory()). It's done to eliminate an initialization exception in
JsonLoader - Jackson Dataformat library does not declare an OSGi
dependency on Jackson Databind which is needed to initialize YAMLMapper
(it extends ObjectMapper from Jackson Databind).