[#75] Added getParent adn getRoot overlay methods
Also, getModel() is a synonym for getRoot() in OpenApiObject, and that
also defines getParentObject, which searches for the nearest ancestor
that's a PropertiesOverlay.

ChildOverlay objects do not participate in the parent hiearchy in
general, but they define their own getParent method that returns this
child's parent - e.g. the holder of the reference if the child is
specified via reference. In such case, the parent of the non-reference
parent, if there is one, can be obtained via
getOverlay().getParent(). (There may not be one, e.g. when an external
JsonNode is parsed due to references in the parsed model, but no
ancestor of that node (in its containing file) is referenced in the
parsed model.