[#75] BigParseTest finally works again!
Lots of fiddling with JsonPath, and had to add some special junk to deal
with additionalProperties.

Also important was preventing map overlays with a key pattern from
participating in the json=>overlay cache in ReferenceRegistry. Such an
overlay only consumes a portion of the JSON object. Two map overlays
built from the same json but with different key patterns would not work,
nor would a map overlay that's built from the same json as properties
overlay (like most of the extensions maps in OAS3).