[#189] New replacement for Messages class
The old one is still there, where it's always been. The new
implementation is in the val.msg package.

The new implemnetation has these improvements:
* No longer includes inline messages, but messages are identified by
  enum values, and it's easy to navigate to the enum definition, where
  the default message appears.
* Properties files can be used to provide localized messages, either for
  a specific two-part language tag (e.g. "en-US") or by a language
  id ("en") as a fallback.
* Multiple enum classes can easily be defined for different sets of
  messages, and they can all be used as needed in the application

This makes it slightly more complicated to compose a new message, but
the resulting code is much more concise, and with an easy click-through
to the default message, this gets around the main issue I've had with
the eclipse org.eclipse.osgi.util.NLS class.